requirements.txt
absl-py==0.7.1
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
future==0.17.1
gast==0.2.2
gpflow==1.3.0
grpcio==1.20.1
gym==0.12.1
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
more-itertools==7.0.0
multipledispatch==0.6.0
numpy==1.16.2
pandas==0.24.2
pbr==5.2.0
pluggy==0.9.0
protobuf==3.7.1
py==1.8.0
pyglet==1.3.2
pyparsing==2.4.0
pytest==4.4.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.2
Werkzeug==0.15.2
